### 2026-01-21 #9581 @amslala
Hi @arfon and @jromanowska! Thank you for the updated information regarding the paper format.

I'll update the paper in the next weeks to fulfill the new requirements.
