### 2026-03-12 #9581 @jromanowska
@amslala - just one small thing: you probably want to have section title "Research impact statement" and not "Research impact statement (200 words)" ;)
And also - check the missing DOIs that editorialbot found and add them, if these are correct.

You can change it and re-generate the pdf with the command:
`@editorialbot generate pdf`
