### 2026-03-12 #9581 @jromanowska
@turkulerc - thank you for agreeing to review this one! We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
I need to find a second reviewer to officially start the process, so for now you can read the docs, take a look at the repo, or just wait :)
