### 2026-03-23 #10240 @editorialbot
[REVIEW]: regkit: facilitating manipulation, analysis and visualization of data from Norwegian health and population registers
**Submitting author:** @amslala (<a href="http://orcid.org/0009-0002-8282-6529">Alejandra Martinez Sanchez</a>)
**Repository:** https://github.com/amslala/regtools
**Branch with paper.md** (empty if default branch):
**Version:** v3.0.1
