### 2026-04-13 #10240 @amslala
> _Contribution and authorship:_ The submitting author [@amslala](https://github.com/amslala) made major contributions to the software with 1 commit from the last author [@lauriejh](https://github.com/lauriejh). The contributions of the remaining c-authors are not visible on the commit history of the repository or on the proof. Hi [@amslala](https://github.com/amslala), would you like to elaborate on this?

@jromanowska and @turkulerc:  Due to the sensitive nature of the data used to construct and test the package, the initial development of the package was carried out within a secure research platform without internet access or version control. As a result, the commits in this repository do not reflect those contributions. All the co-authors contributed conceptually to the iterative development of the package, as well as provided comments on the manuscript in pdf form.
