### 2026-04-16 #10240 @turkulerc
Hi @amslala, I like the [documentation](https://amslala.github.io/regtools/index.html) very much. It is well written and clearly structured, and overall easy to follow. I also appreciate that the functions support the Parquet file format. The section that provides tips about metadata from registries (under “Variables and codes”) is especially valuable, particularly for researchers who are less familiar with data types or unsure where to find metadata. I tested the filtering functions using some invalid ICPC‑2 patterns and received a warning exactly as described in the documentation, which is great.

I did encounter a few minor issues and inconsistencies in the documentation that might be worth addressing:

1) The example dataset _var_df_ has columns "id", "year_varying" and "varying_code", whereas the package help page documents the colums as id, code and year.
2) dataset linked_df includes a diagnosis_count column, which is not documented on the corresponding help page.
3) Setting the argument log_path = log_file in the functions give error, obviously because no objects called log_file has been created prior to running the code. This could be confusing for first-time users with little R experience. It might help to move the part explaining logging to the beginning and/or default log_file to NULL, so the examples can run without modification.
4) Examples that create population files (pop_df) run runif() without setting a seed. Less experienced users may be confused when they see different values. A short note explaining this behavior, or setting a seed in the examples, would improve reproducibility.
5) The "Harmonize municipality codes" example does not work as expected. The resulting table contains many NA values (both [here](https://amslala.github.io/regtools/articles/other-useful-fun.html) and when i ran the code locally). The text states that "The residence_df includes then a column varying_code with the municipality codes in 2017." but the values in varying_code are characters such as 5.1, 5, 3 etc., which are not municipality codes. When I used actual example municipality codes from 2017, the function worked as expected.
6) At the end of the benchmarking section, outputs are shown for mb_filter and autoplot(mb_filter), but these two lines are not included into the code. For completeness, it might be helpful to add these lines at the end of the example code, since the outputs already appear on the [page](https://amslala.github.io/regtools/articles/parquet-files.html).

Regarding dependencies, I ran into a few issues when installing the package or executing the examples:
1) The package PxWebApiData should be installed to run get_population_ssb(), the package ggrepel should be installed to run plot_rates().

I also tested installation in a restricted environment without internet access, where I had to install the package from source. In that setup, I needed to manually install the following dependencies before the package could be installed successfully: binom, ggthemes, klassR, and logger.

As a side note: there are several packages written in R and other languages with name regtools. This name overlap might make discoverability or identification slightly harder for potential users.

Overall, this is a very solid package with high‑quality documentation; the points above are mostly small polish issues that could further improve the user experience.
