### 2026-08-18 #10240 @amslala
Hi @jromanowska. I have completed the points under "Additional Author Tasks":
- Latest release of the package is version v3.0.1
- DOI of release in Zenodo: https://doi.org/10.5281/zenodo.21935587
- The title of the paper and authors are the same in the paper and in the Zenodo archive
- The license is the same in the archive and the software
