### 2026-08-18 #10240 @amslala
Regarding the comments to the paper, see commit [e1672cc](https://github.com/amslala/regkit/commit/e1672cc0102bfa63ad081d916a6b7abb9a344078).

All links in the references should be clickable now. Unfortunatly the report was published without a DOI, but I have referenced to the archived version in NVA.
