### 2026-08-18 #10240 @jromanowska
Thank you, @amslala!
I've noticed you have one more affiliation in the paper, when compared to the zenodo repo, and you need to add the country for each affiliation :)

There is only one link that is not clickable: SSB report from 2023. And maybe when refering to `csverse`, it's better to put [this link](https://www.csids.no/packages.html) or even [the newest page](https://niphr.github.io/)? (I've noticed that the [csverse repos](https://github.com/orgs/csids/repositories?type=source) are no longer maintained)

There is also problem with [doi to your conference paper](https://doi.org/10.7557/5.8265) - it resolves but the webpage doesn't load. I've found [a different reference to it](https://www.semanticscholar.org/paper/Meeting-the-challenges-of-reproducibility-and-in-of-Sanchez-Hannigan/3dfc8f3421285478d80fc17d9d4039059126a0c7), so the DOI is correct, but I'm afraid that the Conference website doesn't hold these papers anymore...?

(After the changes, you don't need to make a new release)
