### 2026-08-18 #10240 @jromanowska
And one more thing - I've noticed the installation instructions tell the user to call `devtools::install_github`, but this function is [deprecated now](https://devtools.r-lib.org/reference/install-deprecated.html). Please, update the docs.
