### 2026-08-21 #10240 @amslala
Hi @jromanowska, I have fixed the affilations in both the paper and zenodo repo.
The link for the SSB report should now be clickable, and I have changed the reference of `csverse` to the new `niphr`.

I think there was a temporary problem with the Septentio webpage that holds the [conference paper](https://doi.org/10.7557/5.8265). It seems like it resolves and loads now.

I have now updated the installation instructions and now use the package `pak`
