### 2026-08-24 #10240 @jromanowska
Thank you for the work, again, @amslala, @turkulerc, @ilovemane - the paper and repo will now be checked for the final time by one of the editors in charge before the official acceptance. :partying_face:
