### 2026-08-24 #10240 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
