### 2026-09-01 #10240 @Kevin-Mattheus-Moerman
@amslala, congratulations on being nearly done :)

As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:

👉  Appears very basic. We do recommend you incorporate dedicated contribution guidelines (e.g. a CONTRIBUTING.md) in the future.

👉 For the moment I see:

I will check again later to see if this is an issue with ZENODO. It should end up saying `Version v3.0.1`.

- [x] Copyedit pass/checked for typos

- [ ] The paper uses mostly American English (e.g. `visualization`, `analyze`, and `harmonizing`) except for `programme` which is European/British English. We would suggest you use one convention consistently.
