### 2026-09-03 #10240 @amslala
@Kevin-Mattheus-Moerman Thank you! I've done the suggested change in the paper, and now it should all be in American English.
